// hohealth-web-app-angular/src/app/user/model/user.model.ts
export class User {
	_id: string | null = null;
	uid: string | null = null;
	name: string | null = null;
	username: string | null = null;
	email: string | null = null;
	phone_number: string | null = null;
	profile_image: string | null = null;
	dob: string | null = null;
	gender: string | null = null;
}